Implemented functionality for rest of buttons and completed GUI

+++ src/models/FileMoveUtil.java
@@ -10,51 +10,61 @@ public class FileMoveUtil extends FileUtil {
     private ArrayList<File> filesCopied;
 
     public void moveFile(String sourcePath, String destPath) {
-	File file = new File(sourcePath);
-	String path = getPath(destPath, file);
-	createDirIfDoesNotExist(destPath);
-	file.renameTo(new File(path));
+		File file = new File(sourcePath);
+		String path = getPath(destPath, file);
+		createDirIfDoesNotExist(destPath);
+		if(!new File(path).exists()) {			
+			file.renameTo(new File(path));
+		}
     }
 
     public void moveFiles(String source, String dest, String query, String searchOption) {
-	setFilesMoved(new ArrayList<File>());
-	for (File file : getFilesInCurrentDir(source)) {
-	    if (file.isDirectory()) {
-		moveFiles(file.getPath(), dest, query, searchOption);
-	    } else if (matches(file.getName(), query, searchOption)) {
-		filesMoved.add(file);
-		moveFile(file.getPath(), dest);
-	    }
-	}
+		setFilesMoved(new ArrayList<File>());
+		for (File file : getFilesInCurrentDir(source)) {
+		    if (file.isDirectory()) {
+		    	moveFiles(file.getPath(), dest, query, searchOption);
+		    } 
+		    else if (matches(file.getName(), query, searchOption)) {
+				filesMoved.add(file);
+				moveFile(file.getPath(), dest);
+		    }
+		}
     }
 
-    public void copyFiles(String source, String dest, String query, String searchOption) throws IOException {
+    public void copyFiles(String source, String dest, String query, String searchOption) {
 	setFilesCopied(filesCopied);
-	for (File file : getFilesInCurrentDir(source)) {
-	    if (file.isDirectory()) {
-		copyFiles(file.getPath(), dest, query, searchOption);
-	    } else if (matches(file.getName(), query, searchOption)) {
-		String path = getPath(dest, file);
-		createDirIfDoesNotExist(dest);
-		File destFile = new File(path);
-		Files.copy(file.toPath(), destFile.toPath());
-	    }
-	}
+		for (File file : getFilesInCurrentDir(source)) {
+		    if (file.isDirectory()) {
+		    	copyFiles(file.getPath(), dest, query, searchOption);
+		    } 
+		    else if (matches(file.getName(), query, searchOption)) {
+				String path = getPath(dest, file);
+				createDirIfDoesNotExist(dest);
+				File destFile = new File(path);
+				try {
+					if(!destFile.exists()) {						
+						Files.copy(file.toPath(), destFile.toPath());
+					}
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+		    }
+		}
     }
 
     public ArrayList<File> getFilesMoved() {
-	return filesMoved;
+    	return filesMoved;
     }
 
     public void setFilesMoved(ArrayList<File> filesMoved) {
-	this.filesMoved = filesMoved;
+    	this.filesMoved = filesMoved;
     }
 
     public ArrayList<File> getFilesCopied() {
-	return filesCopied;
+    	return filesCopied;
     }
 
     public void setFilesCopied(ArrayList<File> filesCopied) {
-	this.filesCopied = filesCopied;
+    	this.filesCopied = filesCopied;
     }
 }
